backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
drf-yasg==1.21.7
django-filter==24.1
django-ckeditor==6.7.2
django-jazzmin==3.0.1
psycopg2-binary==2.9.9
sqlparse==0.5.3
pytz==2024.1
asgiref==3.8.1
typing-extensions==4.11.0
PyYAML==6.0.1
uritemplate==4.1.1
inflection==0.5.1
packaging==24.0
django-js-asset==2.2.0
django-modeltranslation>=0.18.11
googletrans>=4.0.0
djangorestframework-simplejwt==5.3.0
python-telegram-bot==20.7
aiohttp==3.9.1
requests==2.31.0
